+++ examples/service_console/main/audio_service.hpp
@@ -1,8 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2025-2026 Espressif Systems (Shanghai) CO LTD
- *
- * SPDX-License-Identifier: CC0-1.0
- */
-#pragma once
-
-void audio_service_init();